# Fix reflection logic of aws-java-sdk-v2 credential providers